Move stuff into gatsby-shell/
