+++ src/pages/index.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { groupByCategory, toSiteLinks } from '../helpers/site_page'
 import { AllSitePage } from '../types/types'
 import RootPage from '../web/components/RootPage'
-import Layout from '../web/containers/Layout'
+import Layout from '../gatsby-shell/Layout'
 
 /*
  * Types
