+++ src/web/components/RootPage.tsx
@@ -4,7 +4,7 @@ import Helmet from 'react-helmet'
 import useSiteContent from '../../gatsby-hooks/useSiteContent'
 import { GroupedSiteLinks, SiteLink } from '../../types/types'
 import { LiveSearchInput } from '../../web-search'
-import CommonHead from './CommonHead'
+import CommonHead from '../../gatsby-shell/CommonHead'
 import PagesList from './PagesList'
 import SiteHeader from './SiteHeader'
 import TopNav from './TopNav'
